添加vip

+++ ruoyi-xindian/src/main/java/com/ruoyi/xindian/shipAddress/domain/ShipAddress.java
@@ -42,7 +42,7 @@ public class ShipAddress extends BaseEntity {
     /**
      * 删除标记位置
      */
-    private Integer delFlag;
+    private Long delFlag;
 
     public Long getUserId() {
         return userId;
@@ -52,11 +52,11 @@ public class ShipAddress extends BaseEntity {
         this.userId = userId;
     }
 
-    public Integer getDelFlag() {
+    public Long getDelFlag() {
         return delFlag;
     }
 
-    public void setDelFlag(Integer delFlag) {
+    public void setDelFlag(Long delFlag) {
         this.delFlag = delFlag;
     }
